Better npm package, switch to tsup, move types and classes
