--- .npmignore
@@ -1,14 +1,0 @@
-.github/
-.husky/
-coverage/
-**/node_modules
-tests/
-.eslintcache
-.eslintignore
-.eslintrc.cjs
-.gitignore
-.prettierignore
-.prettierrc.cjs
-tsconfig.json
-vite.config.ts
-pnpm-lock.yaml
