+++ package.json
@@ -28,20 +28,24 @@
 	"engines": {
 		"node": ">=16.0.0"
 	},
-	"source": "./src/index.ts",
-	"exports": {
-		"require": "./dist/index.cjs",
-		"default": "./dist/index.modern.js"
-	},
 	"main": "./dist/index.cjs",
 	"module": "./dist/index.module.js",
-	"unpkg": "./dist/index.umd.js",
 	"types": "./dist/index.d.ts",
+	"exports": {
+		".": {
+			"import": "./dist/index.js",
+			"require": "./dist/index.cjs",
+			"types": "./dist/index.d.ts"
+		}
+	},
+	"files": [
+		"dist"
+	],
 	"scripts": {
 		"prepare": "husky install",
 		"prepublish": "pnpm build",
-		"build": "microbundle",
-		"dev": "microbundle watch",
+		"build": "tsup",
+		"dev": "tsup --watch",
 		"test": "vitest",
 		"coverage": "vitest run --coverage",
 		"typecheck": "tsc --noEmit",
@@ -63,9 +67,9 @@
 		"eslint-config-prettier": "^8.5.0",
 		"husky": "^8.0.2",
 		"lint-staged": "^13.1.0",
-		"microbundle": "^0.15.1",
 		"pinst": "^3.0.0",
 		"prettier": "^2.8.1",
+		"tsup": "^6.5.0",
 		"typescript": "^4.9.4",
 		"vitest": "^0.26.2",
 		"vitest-github-actions-reporter": "^0.9.0"
