+++ src/UmamiApiClient.ts
@@ -5,305 +5,46 @@ import axios, {
 	type AxiosRequestConfig,
 	type AxiosResponse,
 } from "axios";
-import { convertPeriodToTime } from "./utils/time-periods";
+import { Website, type WebsiteData } from "./classes/website";
+import { UserAccount, type UserAccountData } from "./classes/user-account";
+import { DEFAULT_HTTP_CLIENT_TIMEOUT_MS, DEFAULT_USER_AGENT } from "./defaults";
 
-const {
-	UMAMI_CLIENT_TIMEOUT_MS,
-	UMAMI_CLIENT_USER_AGENT,
-	UMAMI_CLIENT_TIME_PERIOD,
-	UMAMI_CLIENT_TIME_UNIT,
-	UMAMI_CLIENT_TIMEZONE,
-	UMAMI_CLIENT_METRIC_TYPE,
-} = process.env;
+interface AuthData {
+	token: string;
+	user: {
+		userId: number;
+		username: string;
+		isAdmin: boolean;
+		accountUuid: string;
+		iat?: number;
+		shareToken?: string;
+	};
+}
 
-const DEFAULT_HTTP_CLIENT_TIMEOUT_MS = UMAMI_CLIENT_TIMEOUT_MS
-	? parseInt(UMAMI_CLIENT_TIMEOUT_MS)
-	: 2000;
-const DEFAULT_USER_AGENT =
-	UMAMI_CLIENT_USER_AGENT ??
-	"Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:102.0) Gecko/20100101 Firefox/102.0";
-const DEFAULT_TIME_PERIOD = UMAMI_CLIENT_TIME_PERIOD ?? "24h";
-const DEFAULT_TIME_UNIT = UMAMI_CLIENT_TIME_UNIT ?? "hour";
-const DEFAULT_TIMEZONE = UMAMI_CLIENT_TIMEZONE ?? "America/Toronto";
-const DEFAULT_METRIC_TYPE = UMAMI_CLIENT_METRIC_TYPE ?? "url";
+interface PageViewPayload {
+	website: string;
+	url: string;
+	referrer?: string;
+	hostname: string;
+	language?: string;
+	screen?: string;
+}
+
+interface EventPayload {
+	website: string;
+	url: string;
+	referrer?: string;
+	hostname: string;
+	language?: string;
+	screen?: string;
+	event_name: string;
+	event_data: string;
+}
+
+type CollectPayload = PageViewPayload | EventPayload;
 
 function isAxiosHeaders(headers?: RawAxiosRequestHeaders | AxiosHeaders): headers is AxiosHeaders {
 	return !!headers && !Object.hasOwn(headers, "common");
-}
-
-class Website implements WebsiteData {
-	private readonly _apiClient: UmamiApiClient;
-	private readonly _axios: AxiosInstance;
-	public readonly id: number;
-	public readonly websiteUuid: string;
-	public userId: number;
-	public name: string;
-	public domain: string;
-	public shareId: string | null;
-	public createdAt: string;
-
-	constructor(apiClient: UmamiApiClient, axios: AxiosInstance, data: WebsiteData) {
-		this._apiClient = apiClient;
-		this._axios = axios;
-		this.id = data.id;
-		this.websiteUuid = data.websiteUuid;
-		this.userId = data.userId;
-		this.name = data.name;
-		this.domain = data.domain;
-		this.shareId = data.shareId;
-		this.createdAt = data.createdAt;
-	}
-
-	/**
-	 * Updates the website.
-	 * @param options.domain The domain name of the website (e.g. umami.is)
-	 * @param options.name The name of the website (usually the same as the domain)
-	 * @param options.owner The website's owner's ID (by default, the logged-in's user's)
-	 * @param options.enableShareUrl Whether or not to enable public sharing.
-	 * @returns
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/websites/[id]/index.js#L23-L57 Relevant Umami source code}
-	 */
-	public async update(options: {
-		domain: string;
-		name: string;
-		owner?: number;
-		enableShareUrl?: boolean;
-	}) {
-		if (!options.owner) {
-			const currentUser = await this._apiClient.getCurrentUser();
-			options = {
-				...options,
-				owner: currentUser.userId,
-			};
-		}
-
-		const { data } = await this._axios.post<WebsiteData>(`/websites/${this.websiteUuid}`, options);
-		Object.assign(this, data);
-		return this;
-	}
-
-	/**
-	 * Deletes the website
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/websites/[id]/index.js#L59-L67 Relevant Umami source code}
-	 */
-	public async delete() {
-		await this._axios.delete(`/websites/${this.websiteUuid}`);
-	}
-
-	/**
-	 * Gets the stats of the website from a specified time period
-	 * @param options.period The time period of stats to return
-	 * @param options.url Filter stats by URL
-	 * @param options.referrer Filter stats by referrer
-	 * @param options.os Filter stats by OS
-	 * @param options.browser Filter stats by browser
-	 * @param options.device Filter stats by device
-	 * @param options.country Filter stats by country
-	 * @returns The website's stats from the specified time period
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/websites/[id]/stats.js Relevant Umami source code}
-	 */
-	public async getStats(options?: {
-		period?: TimePeriod;
-		url?: string;
-		referrer?: string;
-		os?: string;
-		browser?: string;
-		device?: string;
-		country?: string;
-	}) {
-		const { data } = await this._axios.get<Stats>(`/websites/${this.websiteUuid}/stats`, {
-			params: {
-				...convertPeriodToTime(options?.period ?? DEFAULT_TIME_PERIOD),
-				url: options?.url,
-				referrer: options?.referrer,
-				os: options?.os,
-				browser: options?.browser,
-				device: options?.device,
-				country: options?.country,
-			},
-		});
-		return data;
-	}
-
-	/**
-	 * Resets the website's stats
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/websites/[id]/reset.js Relevant Umami source code}
-	 */
-	public async resetStats() {
-		await this._axios.post(`/websites/${this.websiteUuid}/reset`);
-		return this;
-	}
-
-	/**
-	 * Gets the pageviews of the website from a specified time period
-	 * @param options.period The time period of pageviews to return
-	 * @param options.unit The interval of time/precision of the returned pageviews
-	 * @param options.tz The timezone you're in (defaults to "America/Toronto")
-	 * @param options.url Filter pageviews by URL
-	 * @param options.referrer Filter pageviews by referrer
-	 * @param options.os Filter pageviews by OS
-	 * @param options.browser Filter pageviews by browser
-	 * @param options.device Filter pageviews by device
-	 * @param options.country Filter pageviews by country
-	 * @returns The website's pageviews from the specified time period
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/websites/[id]/pageviews.js Relevant Umami source code}
-	 */
-	public async getPageviews(options?: {
-		period?: TimePeriod;
-		unit?: TimeUnit;
-		tz?: string;
-		url?: string;
-		referrer?: string;
-		os?: string;
-		browser?: string;
-		device?: string;
-		country?: string;
-	}) {
-		const { data } = await this._axios.get<PageViews>(`/websites/${this.websiteUuid}/pageviews`, {
-			params: {
-				...convertPeriodToTime(options?.period ?? DEFAULT_TIME_PERIOD),
-				unit: options?.unit ?? DEFAULT_TIME_UNIT,
-				tz: options?.tz ?? DEFAULT_TIMEZONE,
-				url: options?.url,
-				referrer: options?.referrer,
-				os: options?.os,
-				browser: options?.browser,
-				device: options?.device,
-				country: options?.country,
-			},
-		});
-		return data;
-	}
-
-	/**
-	 * Gets the events of the website from a specified time period
-	 * @param options.period The time period of events to return
-	 * @param options.unit The interval of time/precision of the returned events
-	 * @param options.tz The timezone you're in (defaults to "America/Toronto")
-	 * @param options.url The url where the event happened.
-	 * @param options.event_type The type of event to request.
-	 * @returns An array of events from the specified time period
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/website/[id]/events.js Relevant Umami source code}
-	 */
-	public async getEvents(options?: {
-		period?: TimePeriod;
-		unit?: TimeUnit;
-		tz?: string;
-		url?: string;
-		event_type?: string;
-	}) {
-		const { data } = await this._axios.get<Event[]>(`/websites/${this.websiteUuid}/events`, {
-			params: {
-				...convertPeriodToTime(options?.period ?? DEFAULT_TIME_PERIOD),
-				unit: options?.unit ?? DEFAULT_TIME_UNIT,
-				tz: options?.tz ?? DEFAULT_TIMEZONE,
-				url: options?.url,
-				event_type: options?.event_type,
-			},
-		});
-		return data;
-	}
-
-	/**
-	 * Gets a type of metrics of the website from a specified time period
-	 * @param options.period The time period of events to return
-	 * @param options.type The type of metric to get. Defaults to url
-	 * @param options.url Filter metrics by URL
-	 * @param options.referrer Filter metrics by referrer
-	 * @param options.os Filter metrics by OS
-	 * @param options.browser Filter metrics by browser
-	 * @param options.device Filter metrics by device
-	 * @param options.country Filter metrics by country
-	 * @returns An array of metrics from the specified time period
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/websites/[id]/metrics.js Relevant Umami source code}
-	 */
-	public async getMetrics(options?: {
-		period?: TimePeriod;
-		type?: MetricType;
-		url?: string;
-		referrer?: string;
-		os?: string;
-		browser?: string;
-		device?: string;
-		country?: string;
-	}) {
-		const { data } = await this._axios.get<Metric[]>(`/websites/${this.websiteUuid}/metrics`, {
-			params: {
-				...convertPeriodToTime(options?.period ?? DEFAULT_TIME_PERIOD),
-				type: options?.type ?? DEFAULT_METRIC_TYPE,
-				url: options?.url,
-				referrer: options?.referrer,
-				os: options?.os,
-				browser: options?.browser,
-				device: options?.device,
-				country: options?.country,
-			},
-		});
-		return data;
-	}
-
-	/**
-	 * Gets the active visitors of a website
-	 * @returns
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/websites/[id]/active.js Relevant Umami source code}
-	 */
-	public async getActiveVisitors() {
-		const { data } = await this._axios.get<ActiveVisitor[]>(`/websites/${this.websiteUuid}/active`);
-		return data;
-	}
-}
-
-class UserAccount implements UserAccountData {
-	private readonly _axios: AxiosInstance;
-	public readonly id: number;
-	public username: string;
-	public isAdmin: boolean;
-	public readonly createdAt: string;
-	public updatedAt: string;
-	public accountUuid: string;
-
-	constructor(axios: AxiosInstance, data: UserAccountData) {
-		this._axios = axios;
-		this.id = data.id;
-		this.username = data.username;
-		this.isAdmin = data.isAdmin;
-		this.createdAt = data.createdAt;
-		this.updatedAt = data.updatedAt;
-		this.accountUuid = data.accountUuid;
-	}
-
-	/**
-	 * Updates a user account
-	 * @param options.username New username (admin only)
-	 * @param options.password New password
-	 * @returns The user account
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/accounts/[id]/index.js#L21-L53 Relevant Umami source code}
-	 */
-	public async update(options: { username: string; password: string }) {
-		const { data } = await this._axios.post(`/accounts/${this.id}`, options);
-		Object.assign(this, data);
-		return this;
-	}
-
-	/**
-	 * Updates a user account password
-	 * @param options.current_password Current password
-	 * @param options.new_password New password
-	 * @returns The user account
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/accounts/[id]/password.js Relevant Umami source code}
-	 */
-	public async changePassword(options: { current_password: string; new_password: string }) {
-		await this._axios.post(`/accounts/${this.accountUuid}/password`, options);
-		return this;
-	}
-
-	/**
-	 * Deletes the user account (admin only)
-	 * @see {@link https://github.com/umami-software/umami/blob/master/pages/api/accounts/[id]/index.js#L55-L63 Relevant Umami source code}
-	 */
-	public async delete() {
-		await this._axios.delete(`/accounts/${this.id}`);
-	}
 }
 
 export default class UmamiApiClient {
